Removed pending ".this" qualifiers Commented some unnused code to skip compiler warning Converted some loops into LinQ Syntax Removed some unnused parameters Cleanup

+++ src/L2dotNET.GameService/network/serverpackets/SpawnItem.cs
@@ -20,7 +20,7 @@ namespace L2dotNET.GameService.Network.Serverpackets
             WriteD(_item.Y);
             WriteD(_item.Z);
             WriteD(_item.Template.Stackable ? 1 : 0);
-            WriteD((int)_item.Count);
+            WriteD(_item.Count);
             WriteD(0); // ?
         }
     }
